Update type and structure
